Extract common helpers.
