+++ packages/utils/src/bigint-uint8-hex-string.ts
@@ -5,37 +5,7 @@
  * environments.
  */
 
-/**
- * Helper for `biToMtcHex()`, which converts a hex digit at a particular index
- * in a string to its 4-bit (nibble-sized) numeric value. Handles '0'-'9'
- * (0x30-0x39) and 'a'-'f' (0x61-0x66).
- */
-function nibbleValueAt(hex: string, at: number): number {
-  const c = hex.charCodeAt(at);
-
-  // '0'-'9' = 0x30-0x39, 'a'-'f' = 0x61-0x66
-  return c < 0x3a ? c - 0x30 : c - 0x57;
-}
-
-/**
- * Helper for `biToMtcHex()`, which converts a positive `bigint` to a hex string
- * with an even number of digits, _and_ a leading `00` if it would otherwise be
- * interpreted as a negative number in twos-complement.
- */
-function hexStringFromPositiveValue(value: bigint): string {
-  const hex = value.toString(16);
-  if ((hex.length & 1) === 1) {
-    // Round up to an even number of nibbles.
-    return "0" + hex;
-  } else if (nibbleValueAt(hex, 0) >= 8) {
-    // Add an extra `0x00` byte, because the high-order bit would otherwise
-    // be `1` and therefore the encoded result would be negative, which
-    // would be wrong.
-    return "00" + hex;
-  } else {
-    return hex;
-  }
-}
+import { hexStringFromPositiveValue } from "./bigint-shared-impl.ts";
 
 /**
  * Version of `bigintToMinimalTwosComplement()` which uses the `Uint8Array`
